requirements.txt
Flask==2.3.3
python-whois==0.8.0
Flask-SQLAlchemy==3.0.5